hugging_face_deployment/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.0.0
numpy>=1.21.0
gradio>=3.0.0
transformers>=4.20.0
huggingface_hub>=0.10.0
